Ajuste al tipo de variable a la puntuacion del SIEVA. y algunos fixes para el css de la aplicacion.

+++ app/DI.Colef.Sia.Core/Articulo.cs
@@ -100,7 +100,7 @@ namespace DecisionesInteligentes.Colef.Sia.Core
 
         public virtual int PaginaFinal { get; set; }
 
-        public virtual int Puntuacion { get; set; }
+        public virtual decimal Puntuacion { get; set; }
 
         public virtual Departamento Departamento { get; set; }
