+++ app/DI.Colef.Sia.Core/Curso.cs
@@ -49,7 +49,7 @@ namespace DecisionesInteligentes.Colef.Sia.Core
 
         public virtual Subdisciplina Subdisciplina { get; set; }
 
-        public virtual int Puntuacion { get; set; }
+        public virtual decimal Puntuacion { get; set; }
 
         [NotNull]
         public virtual Usuario Usuario { get; set; }
