+++ app/DI.Colef.Sia.Core/Libro.cs
@@ -131,7 +131,7 @@ namespace DecisionesInteligentes.Colef.Sia.Core
 
         public virtual RevistaPublicacion RevistaPublicacion { get; set; }
 
-		public virtual int Puntuacion { get; set; }
+		public virtual decimal Puntuacion { get; set; }
 
         [NotNull]
         public virtual Usuario Usuario { get; set; }
