+++ app/DI.Colef.Sia.Web.Controllers/Models/FirmaForm.cs
@@ -27,6 +27,6 @@ namespace DecisionesInteligentes.Colef.Sia.Web.Controllers.Models
         public string Descripcion { get; set; }
         public int ProductoId { get; set; }
 
-        public int PuntuacionSieva { get; set; }
+        public decimal PuntuacionSieva { get; set; }
     }
 }
